pyproject.toml
[project]
name = "network_incident"
version = "0.1"
description = "AI-driven agent designed to facilitate the exploration of the academic landscape"
authors = [{ name = "Antonio Gulli", email = "gulli@google.com" }]
license = "Apache License 2.0"

[tool.poetry]
packages = [
    {include = "./configs"},
    {include = "./backend"}
]


[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.0.0"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = "^1.93"
pandas = "^2.2.3"
ipykernel = "^6.30.1"
elasticsearch = "^9.1.0"
tqdm = "^4.67.1"
langchain-google-vertexai = "^2.0.28"
# configs = { path = "./configs", develop = true }
streamlit = "^1.49.1"
google-cloud-bigquery = "^3.37.0"
google-cloud-storage = "2.18.0"
openpyxl = "^3.1.5"
langchain = "^0.3.27"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { extras = ["eval"], version = "^1.0.0" }
pytest = "^8.3.5"
black = "^25.1.0"
pytest-asyncio = "^0.26.0"
tabulate = "^0.9.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

# [[tool.poetry.packages]]
# include = "configs"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"